Fix OpenStack infra team name
The OpenStack Infrastructure team is more than just OpenStack Foundation
employees. One of the leaders of this team pointed this out in IRC:

<jeblair> in
https://developer.rackspace.com/blog/openstack-innovation-center-updates-january-2016/
it says "The next consumers slated to come on board is the OpenStack
Foundation Infrastructure team."
<jeblair> i appreciate the shout-out, but i wonder if anyone knows
someone at rackspace where we could correct that by removing the word
"foundation", at least in future communications?

This commit fixes that :)